test(research): extract fallback receipt bindings
